examples/virus: Restructure
- VirusAgent in agents.py
- Flatten structure
- Remove old visualisation code
