+++ examples/basic/virus_on_network/model.py
@@ -1,14 +1,10 @@
 import math
-from enum import Enum
 
 import mesa
+from mesa import Model
 import networkx as nx
 
-
-class State(Enum):
-    SUSCEPTIBLE = 0
-    INFECTED = 1
-    RESISTANT = 2
+from agents import VirusAgent, State
 
 
 def number_state(model, state):
@@ -27,7 +23,7 @@ def number_resistant(model):
     return number_state(model, State.RESISTANT)
 
 
-class VirusOnNetwork(mesa.Model):
+class VirusOnNetwork(Model):
     """
     A virus model with some number of agents
     """
@@ -102,65 +98,3 @@ class VirusOnNetwork(mesa.Model):
     def run_model(self, n):
         for i in range(n):
             self.step()
-
-
-class VirusAgent(mesa.Agent):
-    """
-    Individual Agent definition and its properties/interaction methods
-    """
-
-    def __init__(
-        self,
-        model,
-        initial_state,
-        virus_spread_chance,
-        virus_check_frequency,
-        recovery_chance,
-        gain_resistance_chance,
-    ):
-        super().__init__(model)
-
-        self.state = initial_state
-
-        self.virus_spread_chance = virus_spread_chance
-        self.virus_check_frequency = virus_check_frequency
-        self.recovery_chance = recovery_chance
-        self.gain_resistance_chance = gain_resistance_chance
-
-    def try_to_infect_neighbors(self):
-        neighbors_nodes = self.model.grid.get_neighborhood(
-            self.pos, include_center=False
-        )
-        susceptible_neighbors = [
-            agent
-            for agent in self.model.grid.get_cell_list_contents(neighbors_nodes)
-            if agent.state is State.SUSCEPTIBLE
-        ]
-        for a in susceptible_neighbors:
-            if self.random.random() < self.virus_spread_chance:
-                a.state = State.INFECTED
-
-    def try_gain_resistance(self):
-        if self.random.random() < self.gain_resistance_chance:
-            self.state = State.RESISTANT
-
-    def try_remove_infection(self):
-        # Try to remove
-        if self.random.random() < self.recovery_chance:
-            # Success
-            self.state = State.SUSCEPTIBLE
-            self.try_gain_resistance()
-        else:
-            # Failed
-            self.state = State.INFECTED
-
-    def try_check_situation(self):
-        if (self.random.random() < self.virus_check_frequency) and (
-            self.state is State.INFECTED
-        ):
-            self.try_remove_infection()
-
-    def step(self):
-        if self.state is State.INFECTED:
-            self.try_to_infect_neighbors()
-        self.try_check_situation()
