--- examples/basic/virus_on_network/requirements.txt
@@ -1,2 +1,0 @@
-networkx>=2.0
-mesa~=2.0
