--- examples/basic/virus_on_network/run.py
@@ -1,3 +1,0 @@
-from virus_on_network.server import server
-
-server.launch(open_browser=True)
